+++ nanohub/src/routes/upload/-utils.ts
@@ -1,4 +1,4 @@
-import { isTextContentType, TEXT_FILE_EXTENSION_SET } from 'clawhub-schema'
+import { isTextContentType, TEXT_FILE_EXTENSION_SET } from 'nanohub-schema'
 import { getUserFacingConvexError } from '../../lib/convexError'
 
 export async function uploadFile(uploadUrl: string, file: File) {